+++ infra/common/src/main/java/org/apache/shardingsphere/infra/metadata/database/schema/reviser/table/TableMetaDataReviseEngine.java
@@ -25,6 +25,8 @@ import org.apache.shardingsphere.infra.metadata.database.schema.reviser.constrai
 import org.apache.shardingsphere.infra.metadata.database.schema.reviser.index.IndexReviseEngine;
 import org.apache.shardingsphere.infra.rule.ShardingSphereRule;
 
+import java.util.Collection;
+import java.util.Collections;
 import java.util.Optional;
 
 /**
@@ -46,10 +48,35 @@ public final class TableMetaDataReviseEngine<T extends ShardingSphereRule> {
      * @return revised table meta data
      */
     public TableMetaData revise(final TableMetaData originalMetaData) {
+        return revise(originalMetaData, Collections.singleton(originalMetaData));
+    }
+    
+    /**
+     * Revise table meta data.
+     *
+     * @param originalMetaData original table meta data
+     * @param originalMetaDataList original table meta data list
+     * @return revised table meta data
+     */
+    public TableMetaData revise(final TableMetaData originalMetaData, final Collection<TableMetaData> originalMetaDataList) {
+        return revise(originalMetaData, originalMetaDataList, Collections.emptyList());
+    }
+    
+    /**
+     * Revise table meta data.
+     *
+     * @param originalMetaData original table meta data
+     * @param originalMetaDataList original table meta data list
+     * @param schemaMetaDataRevisionCandidateTableMetaDataList schema meta data revision candidate table meta data list
+     * @return revised table meta data
+     */
+    public TableMetaData revise(final TableMetaData originalMetaData, final Collection<TableMetaData> originalMetaDataList,
+                                final Collection<TableMetaData> schemaMetaDataRevisionCandidateTableMetaDataList) {
         Optional<? extends TableNameReviser<T>> tableNameReviser = reviseEntry.getTableNameReviser();
         String revisedTableName = tableNameReviser.map(optional -> optional.revise(originalMetaData.getName(), rule)).orElse(originalMetaData.getName());
         return new TableMetaData(revisedTableName, new ColumnReviseEngine<>(rule, reviseEntry).revise(originalMetaData.getName(), originalMetaData.getColumns()),
-                new IndexReviseEngine<>(rule, reviseEntry).revise(originalMetaData.getName(), originalMetaData.getIndexes()),
+                new IndexReviseEngine<>(rule, reviseEntry).revise(originalMetaData.getName(), originalMetaData.getIndexes(), originalMetaDataList,
+                        schemaMetaDataRevisionCandidateTableMetaDataList),
                 new ConstraintReviseEngine<>(rule, reviseEntry).revise(originalMetaData.getName(), originalMetaData.getConstraints()), originalMetaData.getType());
     }
 }